+++ FamilyTreeFIR/FamilyTree/扩展方法/NSString+arabicToChinese.m
@@ -9,7 +9,7 @@
 #import "NSString+arabicToChinese.h"
 
 @implementation NSString (arabicToChinese)
-+(NSString *)translation:(int)arebic{
++(NSString *)translation:(int64_t)arebic{
     NSString *str = [NSString stringWithFormat:@"%d",arebic];
     NSArray *arabic_numerals = @[@"1",@"2",@"3",@"4",@"5",@"6",@"7",@"8",@"9",@"0"];
     NSArray *chinese_numerals = @[@"一",@"二",@"三",@"四",@"五",@"六",@"七",@"八",@"九",@"零"];
@@ -41,7 +41,6 @@
                 continue;
             }
         }
-        
         [sums addObject:sum];
     }
